[#22] ProductItem 코드 엔티티 파일로 분기
