+++ Codive/Features/Add/Domain/Entities/ProductItem.swift
@@ -1,0 +1,17 @@
+//
+//  ProductItem.swift
+//  Codive
+//
+//  Created by 황상환 on 11/18/25.
+//
+
+import Foundation
+
+// MARK: - Product Item Model
+struct ProductItem: Identifiable {
+    let id = UUID()
+    let imageName: String
+    let isTodayCloth: Bool
+    let brand: String?
+    let name: String?
+}
